+++ lib/widget/order/sales_order_detail.dart
@@ -2,9 +2,12 @@
 import "package:flutter/material.dart";
 import "package:flutter_speed_dial/flutter_speed_dial.dart";
 import "package:font_awesome_flutter/font_awesome_flutter.dart";
+import "package:inventree/barcode/barcode.dart";
+import "package:inventree/barcode/sales_order.dart";
 import "package:inventree/inventree/company.dart";
 import "package:inventree/inventree/sales_order.dart";
 import "package:inventree/widget/order/so_line_list.dart";
+import "package:inventree/widget/order/so_shipment_list.dart";
 import "package:inventree/widget/refreshable_state.dart";
 
 import "package:inventree/l10.dart";
@@ -37,7 +40,6 @@ class _SalesOrderDetailState extends RefreshableState<SalesOrderDetailWidget> {
   List<InvenTreeSOLineItem> lines = [];
 
   bool supportsProjectCodes = false;
-  int completedLines = 0;
   int attachmentCount = 0;
 
   @override
@@ -59,6 +61,25 @@ class _SalesOrderDetailState extends RefreshableState<SalesOrderDetailWidget> {
     }
 
     return actions;
+  }
+
+  // Add a new shipment against this sales order
+  Future<void> _addShipment(BuildContext context) async {
+
+    var fields = InvenTreeSalesOrderShipment().formFields();
+
+    fields["order"]?["value"] = widget.order.pk;
+    fields["order"]?["hidden"] = true;
+
+    InvenTreeSalesOrderShipment().createForm(
+      context,
+      L10().shipmentAdd,
+      fields: fields,
+      onSuccess: (result) async {
+        refresh(context);
+      }
+    );
+
   }
 
   // Add a new line item to this sales order
@@ -93,6 +114,16 @@ class _SalesOrderDetailState extends RefreshableState<SalesOrderDetailWidget> {
           }
         )
       );
+
+      actions.add(
+        SpeedDialChild(
+          child: FaIcon(FontAwesomeIcons.circlePlus),
+          label: L10().shipmentAdd,
+          onTap: () async {
+            _addShipment(context);
+          }
+        )
+      );
     }
 
     return actions;
@@ -102,7 +133,37 @@ class _SalesOrderDetailState extends RefreshableState<SalesOrderDetailWidget> {
   List<SpeedDialChild> barcodeButtons(BuildContext context) {
     List<SpeedDialChild> actions = [];
 
-    // TODO
+    if (widget.order.isOpen && InvenTreeSOLineItem().canCreate) {
+      actions.add(
+        SpeedDialChild(
+          child: Icon(Icons.barcode_reader),
+          label: L10().lineItemAdd,
+          onTap: () async {
+            scanBarcode(
+              context,
+              handler: SOAddItemBarcodeHandler(salesOrder: widget.order),
+            );
+          }
+        )
+      );
+
+      if (api.supportsBarcodeSOAllocateEndpoint) {
+        actions.add(
+          SpeedDialChild(
+            child: FaIcon(FontAwesomeIcons.rightToBracket),
+            label: L10().allocateStock,
+            onTap: () async {
+              scanBarcode(
+                context,
+                handler: SOAllocateStockHandler(
+                  salesOrder: widget.order,
+                )
+              );
+            }
+          )
+        );
+      }
+    }
 
     return actions;
   }
@@ -113,14 +174,6 @@ class _SalesOrderDetailState extends RefreshableState<SalesOrderDetailWidget> {
     await api.SalesOrderStatus.load();
 
     supportsProjectCodes = api.supportsProjectCodes && await api.getGlobalBooleanSetting("PROJECT_CODES_ENABLED");
-
-    completedLines = 0;
-
-    for (var line in lines) {
-      if (line.isComplete) {
-        completedLines += 1;
-      }
-    }
 
     InvenTreeSalesOrderAttachment().count(filters: {
       "order": widget.order.pk.toString()
@@ -219,16 +272,16 @@ class _SalesOrderDetailState extends RefreshableState<SalesOrderDetailWidget> {
       ));
     }
 
-    Color lineColor = completedLines < widget.order.lineItemCount ? COLOR_WARNING : COLOR_SUCCESS;
+    Color lineColor = widget.order.complete ? COLOR_SUCCESS : COLOR_WARNING;
 
     tiles.add(ListTile(
       title: Text(L10().lineItems),
       subtitle: ProgressBar(
-        completedLines.toDouble(),
+        widget.order.completedLineItemCount.toDouble(),
         maximum: widget.order.lineItemCount.toDouble()
       ),
       leading: FaIcon(FontAwesomeIcons.clipboardCheck),
-      trailing: Text("${completedLines} / ${widget.order.lineItemCount}", style: TextStyle(color: lineColor)),
+      trailing: Text("${widget.order.completedLineItemCount} / ${widget.order.lineItemCount}", style: TextStyle(color: lineColor)),
     ));
 
     // TODO: total price
@@ -286,8 +339,7 @@ class _SalesOrderDetailState extends RefreshableState<SalesOrderDetailWidget> {
     return [
       Tab(text: L10().details),
       Tab(text: L10().lineItems),
-      // TODO: Add in the "shipped items" tab
-      // Tab(text: L10().shipped)
+      Tab(text: L10().shipments),
     ];
   }
 
@@ -296,7 +348,7 @@ class _SalesOrderDetailState extends RefreshableState<SalesOrderDetailWidget> {
     return [
       ListView(children: orderTiles(context)),
       PaginatedSOLineList({"order": widget.order.pk.toString()}),
-      // Center(), // TODO: Delivered stock
+      PaginatedSOShipmentList({"order": widget.order.pk.toString()}),
     ];
   }